Fix bug when using parameter for field-width.
Caching Replicate.reps_value causes the parameterized field-width to be
"locked" to the value of the parameter that is used for the first
instantiation of some component type.

Then all subsequent instantiations reuse that value, even if a different
parameter value was specified in the code.

The fix is to not cache the reps_value.